Fix #62: Unify LayoutEditorView and OrderPrintDo with Template Method
